notification/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Telethon==1.34.0
nest-asyncio==1.6.0 